:= fix refresh view + handle tasks of other users

+++ top_threads.py
@@ -3,6 +3,7 @@
 import argparse
 import curses
 import datetime
+import errno
 import logging
 import os
 import re
@@ -126,11 +127,30 @@ def get_systat_version():
 
 
 def check_pid(pid):
-    """ Check For the existence of a unix pid. """
+    """Check whether pid exists in the current process table.
+    UNIX only.
+    """
+    if pid < 0:
+        return False
+    if pid == 0:
+        # According to "man 2 kill" PID 0 refers to every process
+        # in the process group of the calling process.
+        # On certain systems 0 is a valid PID but we have no way
+        # to know that in a portable fashion.
+        raise ValueError('invalid PID 0')
     try:
         os.kill(pid, 0)
-    except OSError:
-        return False
+    except OSError as err:
+        if err.errno == errno.ESRCH:
+            # ESRCH == No such process
+            return False
+        elif err.errno == errno.EPERM:
+            # EPERM clearly means there's a process to deny access to
+            return True
+        else:
+            # According to "man 2 kill" possible error values are
+            # (EINVAL, EPERM, ESRCH)
+            raise
     else:
         return True
 
@@ -490,6 +510,9 @@ class StatsRefreshPrinter:
 
     @staticmethod
     def prepare_lines(lines, max_columns):
+        for line in lines:
+            for chunk in line:
+                chunk.text = chunk.text.replace('\t', ' ' * 4)
         new_lines = list(map(lambda line: StatsRefreshPrinter.prepare_line(line, max_columns), lines))
         return reduce(lambda x, y: x + y, new_lines) if len(new_lines) > 1 else new_lines
